feat: Add selector to retrieve university ratings
This commit adds the `uniRatingSelector` to the `recoil/selectors` directory. The selector is responsible for fetching the ratings of universities from the backend API. It uses the `axios` library to make a GET request to the `/uni/get-rating` endpoint, passing the user's authorization token in the headers. If successful, the ratings data is returned. If there is an error, the appropriate error message is thrown.

This commit addresses the need to retrieve university ratings and provides a reusable selector for this purpose.